implement latest storage api changes

+++ esphome/components/binary_storage/i2c_eeprom.h
@@ -76,9 +76,9 @@ class I2CEeprom : public BinaryStorage, public i2c::I2CDevice {
   uint32_t get_page_size() const override { return this->page_size_; }
 
   // RawStorage interface
-  storage::StorageError read(size_t offset, uint8_t *buf, size_t len, size_t *bytes_transferred) override;
-  storage::StorageError write(size_t offset, const uint8_t *buf, size_t len, size_t *bytes_transferred) override;
-  storage::StorageError erase(size_t offset, size_t len) override;
+  storage::StorageError read(uint64_t offset, uint8_t *buf, size_t len, size_t *bytes_transferred) override;
+  storage::StorageError write(uint64_t offset, const uint8_t *buf, size_t len, size_t *bytes_transferred) override;
+  storage::StorageError erase(uint64_t offset, size_t len) override;
   storage::StorageError format() override;
 
   // Hardware-level byte access
