+++ esphome/components/binary_storage/spi_mram.h
@@ -68,9 +68,9 @@ class SPIMRAM : public BinaryStorage,
   uint32_t get_page_size() const override { return 1; }  // No page limit
 
   // RawStorage interface
-  storage::StorageError read(size_t offset, uint8_t *buf, size_t len, size_t *bytes_transferred) override;
-  storage::StorageError write(size_t offset, const uint8_t *buf, size_t len, size_t *bytes_transferred) override;
-  storage::StorageError erase(size_t offset, size_t len) override;
+  storage::StorageError read(uint64_t offset, uint8_t *buf, size_t len, size_t *bytes_transferred) override;
+  storage::StorageError write(uint64_t offset, const uint8_t *buf, size_t len, size_t *bytes_transferred) override;
+  storage::StorageError erase(uint64_t offset, size_t len) override;
   storage::StorageError format() override;
 
   // Hardware-level byte access
